Improve edm command line interface
Bump version to 0.3.0

Move checking out of revs/tags to its own function

Add current rev of repo to git info and show it in detached HEAD state

Add a way to create a snapshot of the current repositories
This scans the current working_dir for git repositories and writes a snapshot.yaml with an entry for the current git rev

IF available in the config checkout the specific git_rev of a repo

Restructure git info gathering

Add edm init command to intialize a workspace

You can either provide a config with a command line argument or the most recent everest-complete config will be downloaded from the everest-dev-environment github repo.
To make this work a new dependency for the requests library is added

Add the edm list command to list available workspaces

Add the edm git info command

This command lists information about git repos in the workspace

Add the edm git pull command.

This command allows you to git pull repos in the current workspace

Add edm snapshot command

Use readonly config if no ssh public key was uploaded to GitHub

Added additional warning/info if no SSH key was found

Use global workspace config when available during cmake mode

Add edm rm command to remove a workspace from the config.

This command only modifies the config and will not delete any files

If a dependency uses a branch as git_tag, use the HEAD rev for CPM

This should alleviate a lot of the problems with broken caches

Remove cmake files, this is now handled by everest-cmake

Add everest-cmake to everest-complete configs

Generated new bash completion file

Signed-off-by: Kai-Uwe Hermann <kai-uwe.hermann@pionix.de>