Use versionsort.suffix when listing tag names (fixes #39)
Tag names including suffixes like -rcXY or similar needs to be
sorted before the final release/tag.

This is what git's versionsort.suffix is intended for, so let's use it.

Let's use a wide matching here, so that -rcX, -pre, -alpha etc.
are covered - just to be on the safe side.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>
Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>